requirements.txt
tensorflow
pandas
numpy
scikit-learn
tensorboard
matplotlib
streamlit
ipykernel

mysql-connector-python
uvicorn